Created a common scan function
I removed code that was redundant between Zones 2 and 3 and created a common scan function.
